Specify image to use in task yaml
In place of specifying the public docker image to use for the compile
jobs with in the pipeline itself, we now will specify the image_resource
to use directly within the task yaml itself.

This means we will now have a task yaml per image resource (per OS).

This change is necessary for migrating to concourse 3.X.X because of
changes to image resources and image_resource. This is not a pattern we
intend to keep indefinitely, it is an MVP pattern just to allow the
migration.

Signed-off-by: Divya Bhargov <dbhargov@pivotal.io>